backend/requirements.txt
# Flask 核心
flask==3.1.2
flask-cors==6.0.2

# AI/ML（已验证兼容版本 - 2024-12-16）
# 注意：这些版本经过测试，兼容 CUDA 12.1 和 Quadro RTX 5000
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
sentence-transformers==2.5.0
accelerate==0.26.1
peft==0.7.1
auto-gptq==0.7.1

# 向量数据库
chromadb==0.4.22

# 数据处理（必须使用 numpy 1.x）
numpy==1.26.4
pandas==2.2.0

# 文件解析
pdfminer.six==20221105
python-docx==1.1.0
openpyxl==3.1.2

# 工具
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
